requirements.txt
# Core dependencies
python-telegram-bot==20.3
aiocron==1.8
openpyxl==3.1.2
pytz==2023.3
python-dotenv==1.1.0
httpx==0.24.1  # Добавлено явное указание версии httpx
python-dotenv>=0.19.0

# Optional for colored logs (Windows)
colorlog==6.7.0

# Ensure pip version (optional)
pip>=25.0.1
